Move utility file into utils dir
